software-map: move the ELK/libavoid layout engine to c4-layout-geometry
Agent-Session: 01a067ce-34b7-7952-a023-ced44207d9ca
